docs: Revise README.md for Kaito inference (#581)
This PR revised the README.md of Kaito inference with more details about
APIs, usages, etc. A workload section is added to explain how we
construct the pod for inference with adapters. The troubleshooting
section is left empty for now.

This change also fixes the wrong sku name in the tuning README.md.

---------

Signed-off-by: Fei Guo <vrgf2003@gmail.com>
Co-authored-by: Ishaan Sehgal <ishaanforthewin@gmail.com>